The Search View While we're here, go type a few letters into that search box. Yep! This is the search view. It's almost identical to list: it re-uses its template and has almost identical config keys.
Overriding Entity View Config

In addition to title, one other key that every view has is help. First, set this below the Genus section: 'Genuses are not covered by warranty! '
Notice, this is directly under the entity config, not under a specific view. Thanks to this, it will apply to all views for this entity. And... yep! It's at the top of the search page, on list and on edit.