In the spirit of EasyAdminBundle, you can override this for each view below. For list, set help to Do not feed!
And the search view still uses the default message. If you want to turn off the help message here, you can totally do that. Under search, set help to null

Refresh! Gone!

Ok, so we're not going to talk about all of the different keys available: it's all pretty easy. The most important thing is to realize that each of the 5 views can be configured on two different levels.

Next, let's talk about actions and the different ones that we can enable.