Adding __toString() Methods In a lot of places, EasyAdminBundle tries to convert your entity objects into strings... ya know, so it can print them in a list or create a drop-down menu. To get this working, we need to add __toString() methods to each entity. Let's do that real quick! I'm casting this into a string just to be safe - if it's null, you'll get an error. In GenusNote, do the same thing.
In GenusScientist, we can return {dollar_sign}this->getUser(). That's an object... but we're about to add a __toString() for it too!

SubFamily, well hey! It already has a __toString(). I'll just add the string cast.

And finally, in User ... make this a bit fancier. Return {dollar_sign}this->getFullName() or {dollar_sign}this->getEmail(), in case the user doesn't have a first or last name in the database.

Try the show page again! Nice! It renders all of the properties, including the relations, which is why it needed that __toString() method.

And because we have a SubFamily admin section, the SubFamily is a link that takes us to its show view.