Removing Actions Speaking of SubFamily, as you can see... there's not much to it: just an id and a name. And the show view, well, it's just the id and name again. Not too interesting. In this case, I think it's overkill to have the show action for the SubFamily entity. So let's kill it!
Back in config.yml, we just added the, show action to the list view globally. Now, under SubFamily, we can override that list config. Add actions and - to remove an action - use -show. Yep, use 'minus' to take an action away.

Refresh! Ah, ha! The show link is gone.

But... even though the link is gone, you can totally still get to the show page! For example, if we click, 'Edit', we can be annoying and change the action in the URL to show. Genus!

Or... you can just go to the show page for any Genus: there is still a link to the SubFamily show page.