Disabling Actions But... even though the link was gone, you could totally still get to the show page! For example, if we clicked, 'Edit', we could be annoying and changed the action in the URL to show. Genus!
Or... you could just go to the show page for any Genus: there was still a link to the SubFamily show page.

That might be ok, but if you truly want to disable the show view, there's a special config key for that. Under the entity itself, add disabled_actions set to an array with show inside.

As soon as we do that ... the link vanishes in dramatic fashion! Let's be annoying again: I'll go forward in my browser to get back to the show page URL. Refresh! Dang! Now we get a huge error. The show view is totally gone.