Customizing Action Design There are two more things you can do with actions: custom actions - we'll talk about those later - and making your actions pretty. That's probably even more important!
I want to tweak how the list actions look... but only for the Genus section. Ok, find its config, go under list and add actions.

This time, rather than adding or removing actions, we want to customize them. So instead of using a simple string like show, use an expanded configuration with name: edit. We are now proudly configuring the edit action.

There are a few things we can do here, like icon: pencil and label: Edit

The icon option - which shows up in a few places - allows you to add Font Awesome icons. For the value, just use whatever comes after the fa-. So, to get the fa-pencil icon, say pencil.

Make the show action just as fancy, with icon: info-circle and a blank label

OoooOooo. Refresh to see how that looks!

Oh man, it's actually kind of ugly... I need to work on my styling skills. But it totally works!