We've been tweaking a bunch of stuff on the list view. But... what about this big giant table in the middle! How can we customize that?
Actually, EasyAdminBundle did a pretty good job with it: it guesses which fields to show, humanizes the column header labels and renders things nicely.

The EasyAdminBundle Profiler

Before we tweak all of this, see that magic wand in the web debug toolbar? Click to open that in a new tab. This is the EasyAdminBundle profiler... and it's awesome. Here, under 'Current entity configuration', we can see all of the config we've been building for this entity, including default values that it's guessing for us. This is a sweet map for knowing what can be changed and how.

Under list, then fields, it shows details about all the columns used for the table. For example, under name, you can see type => string. Actually, dataType is the really important one.

Here's the deal: each field that's rendered in the table has a different data type, like string, float, date, email and a bunch others. EasyAdminBundle guesses a type, and it affects how the data for that field is rendered. We can change the type... and anything else you see here.