This repository Search Pull requests Issues Marketplace Explore  @petre-symfony  Sign out  Watch 0   Star 0   Fork 0 petre-symfony/EasyAdminBundle-KNP-tutorial  Code  Issues 0  Pull requests 0  Projects 0  Wiki  Settings Insights Browse files Controller Fields How? Under Genus and list, add fields. Now, list the exact fields that you want to display, like id, name and isPublished.
These 3 fields were already shown before.

Let's also show firstDiscoveredAt... but! I want to tweak it a little. Just like with actions, there is an 'expanded' config format. Add { } with property: firstDiscoveredAt.

Now... what configuration can we put here? Because this is a date field, it has a format option. Set it to M Y. And, all fields have a label option. Use 'Discovered'

Keep going! Add funFact and then one more expanded property: property: speciesCount. This is an integer type, which also has a format option. For fun, set it to %b - binary format!

Yea know because, scientists are nerds and like puzzles.

 Tip
The format option for number fields is passed to the sprintf() function.

If your head is starting to spin with all of these types and options that I'm pulling out of the air, don't worry! Almost all of the options - like label - are shared across all the types. There are very few type-specific options like format.

And more importantly, in a few minutes, we'll look at a list of all of the valid types and their options.

Ok! Close the profiler tab and refresh. Bam! The table has our 6 columns!

Try out the search again: look for 'quo'. Ok nice! Without any work, the search view re-uses the fields config from list.