Navigate to the User entity section... open up the EasyAdminBundle profiler... and check out the list fields.
I want to talk a bit more about these field 'data types'. Check out isScientist. Its data type is set to toggle, my favorite type! Go back to the main page of the documentation and open List, Search and Show Views Configuration(http://symfony.com/doc/current/bundles/EasyAdminBundle/book/list-search-show-configuration.html).

List Field Types and Options

Down the page a little, it talks about how to 'Customize the Properties Appearance'. This is great stuff. First, it lists the valid options for all the fields, like property, label, css_class, template - which we'll talk about later - and yes! The type, which controls that dataType config. There are a bunch of built-in types, including all of the Doctrine field types and a few special fancy ones from EasyAdminBundle, like toggle. The 'toggle' type is actually super cool: it renders the field as a little switch that turns the value on and off via AJAX.