DQL Filtering Ok, let's turn to something fun: DQL filtering. Like, what if we want to hide some genuses entirely from the list and search page?
But first, so far, it seems like we're limited to one entity section per entity. That's a lie! Let me show you: add a new section under entities called GenusHorde - I just made that up. Below, set its class to AppBundle\Entity\Genus

You see, some scientists are worried that certain genuses are becoming too large... and threaten the survival of mankind. They want a new GenusHorde section where they can keep track of all of the genuses that have a lot of species. It's scary stuff, so we'll add a label: HORDE of Genuses with a scary icon.

Tip
Fun fact! You can press Control+Command+Space to open up the icon menu on a Mac.

And all of a sudden... ah! We have a new 'Horde of Genuses' section! Run!
Of course, this still shows all genuses. I want to filter this to only list genuses that have a lot of species.