Ok, first challenge! I want to override the way the id field is rendered for Genus: add a little key icon next to the number... ya know, because it's the primary key.
This means we need to override the field_id.html.twig template, because id is actually a data type. Copy field_id.html.twig. Then, in app/Resources/views, I already have an admin directory. So inside that, create a new fields directory and paste the file there, as _id.html.twig. Now, add the icon: fa fa-key.

Cool! I put the file here... just because I already have an admin directory. But EasyAdminBundle doesn't automagically know it's there. Nope, we need to tell it. In config.yml, to use this only for Genus, add a templates key, then field_id - the name of the original template - set to admin/fields/_id.html.twig

Try that! Yes! It is using our template... and only in the Genus section.