You may have noticed why everything was gone. We did change the _list_item_actions variable... but we forgot to call the parent block. Add {{ parent() }}.
Try it again. Got it! The delete icon is only there when the item is not published. This was a tricky example... which is why we did it! But usually, customizing things is easier. Technically, the user could still go directly to the URL to delete the Genus, but we'll see how to close that down later.