The subject property is actually an array. But, it has a class key set to the User class. We can use that to make sure we only run our code when we're editing a user.
In other words, {dollar_sign}config = {dollar_sign}event->getSubject() and if {dollar_sign}config['class'] is equal to our User class, then we want to check security. Let's call a new method... that we'll create in a moment: {dollar_sign}this->denyAccessUnlessSuperAdmin(). At the bottom, add that: private function denyAccessUnlessSuperAdmin(). Now... we just need to check to see if the current user has ROLE_SUPERADMIN. How? Via the 'authorization checker' service. To get it, type-hint a new argument with AuthorizationCheckerInterface. Hit Alt+Enter to create and set that property. Then, back down below, if (! {dollar_sign}this->authorizationChecker->isGranted('ROLE_SUPERADMIN'), then throw a new AccessDeniedException(). Make sure you use the class from the Security component. Oh, and don't forget the new!
See, normally, in a controller, we call {dollar_sign}this->denyAccessUnlessGranted(). When we do that, this is actually the exception that is thrown behind the scenes. In other words, we're really doing the same thing that we normally do in a controller.

And... we're done! The service is set to be autowired, so Symfony will know to pass us the authorization checker automatically. Refresh!

Great news! Access denied!