Hiding the Edit Button Time for step 2! On the list page, we need to hide the edit link, unless I have the role. This is trickier: there's no official hook inside of EasyAdminBundle to conditionally hide or show actions. But don't worry! Earlier, we overrode the list template so that we could control exactly what actions are displayed. Our new filter_admin_actions() filter lives in EasyAdminExtension. And we added logic there to hide the delete action for any published genuses. In other words, we added our own hook to control which actions are displayed. We rock!
To hide the edit action, we'll need the authorization checker again. No problem! Add public function __construct() with one argument: AuthorizationCheckerInterface. Set that on a new property. Then, down below, we'll add some familiar code: if  instanceof User and ! {dollar_sign}this->authorizationChecker->isGranted('ROLE_SUPERADMIN'), then unset the edit action.

Phew! It's not the easiest thing ever... EasyAdminBundle... but it does get the job done!

Except for one... minor problem... there is also an edit button on the show page. Oh no! It looks like we need to repeat all of this for the show template!