We got this question: How can I add a 'CSV Export' button to my admin list pages? And now we know why you asked this question: it's tricky! I want to add a button on top that says 'Export'. But, when you add a custom action to the list page... those create links next to each item, not on top. That's not what we want!
Defining the Custom Action

So let's see if we can figure this out. First, in config.yml, under the global list, we are going to add a new action, called export.

Now, if we refresh... not surprisingly, this adds an 'export' link next to every item. And if we click it, it tries to execute a new exportAction() method.

So this is a bit weird: we do not want this new link on each row - we'll fix that in a few minutes. But we do need the new export action. Why? Because as soon as we add this, it's now legal to create a link that executes exportAction(). And that means that we could manually add this link somewhere else... like on top of the list page.